requirements.txt
# 核心推理框架
torch>=2.1.0
deepspeed>=0.12.0
vllm>=0.2.0
transformers>=4.35.0

# API服务框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# 多模态处理
tokenizers>=0.15.0
Pillow>=10.0.0
numpy>=1.24.0

# 分布式通信
ray[serve]>=2.8.0
torch-distributed>=0.1.0

# 性能优化
triton>=2.1.0
flash-attn>=2.3.0
xformers>=0.0.22

# 量化支持
auto-gptq>=0.5.0
awq>=0.1.0
bitsandbytes>=0.41.0

# 监控和日志
prometheus-client>=0.19.0
psutil>=5.9.0
nvidia-ml-py>=12.535.0
wandb>=0.16.0

# 配置管理
pyyaml>=6.0
hydra-core>=1.3.0
omegaconf>=2.3.0

# 工具库
tqdm>=4.66.0
rich>=13.7.0
typer>=0.9.0

# 开发工具
pytest>=7.4.0
black>=23.11.0
isort>=5.12.0
mypy>=1.7.0
